Separate wp-core from wp-content application logic.
The customized sample config.php extends the wp-config-supercharged package to
fit projects where the wp-content folder has been moved outside of the core wp folder.

This technique is useful, for instance, when builduing a composer package with a vcs controlled project:
it avoids saving the wordpress core files as part of the package files.

The bootstrap allows handling Linux Debian/Raspberian environments slightly differently in order to
take the relevant differences into account.